requirements.txt
asgiref==3.8.1
attrs==24.2.0
coverage==7.6.7
Django==5.1
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
drf-spectacular==0.27.2
exceptiongroup==1.2.2
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
pluggy==1.5.0
pytest==8.3.3
pytest-django==4.9.0
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.21.0
sqlparse==0.5.1
tomli==2.1.0
typing_extensions==4.12.2
uritemplate==4.1.1
